UX: Set a delay after confirming 2FA token before redirecting
Before this commit, the user was redirected immediately after confirming
their 2FA token. This means that the success message may not have been
seen by the user completely before the transition happens. This is
alittle adrupt so we want to introduce some delay for the success
message to be shown and read before redirecting.