UX: align the previewless composer with posts (#31960)
This aligns the composer with the post when the preview is
hidden. 

Once the width gets a bit too narrow it occupies the full width. 

[Screen Recording 2025-03-21 at 2.50.50
PM.webm](https://github.com/user-attachments/assets/4a854642-5b20-41fd-a478-7a943dfe0dd4)

The sidebar requires this to get a little magic numbery, but I tried to
anchor to predictable values otherwise.

---------

Co-authored-by: chapoi <101828855+chapoi@users.noreply.github.com>