DEV: Bump Capybara.default_max_wait_time to 20 on CI
There are flaky system tests which have been exhausting the 10 seconds
default max wait time which we have set previously on CI. However, I
have yet to be able to figure out why and lack the tools to be able to
figure out why. The hope here is that the upcoming playwright driver
will provide us with better tooling to debug such problems.

For now, we hope to reduce the occurences of those flaky tests just by
bumping up the default max wait time and see if the occurence of flaky
tests decreases.