FIX: Prioritize the author when replying to topic
Added in https://github.com/discourse/discourse/pull/32086 this prioritization did not accounted for when the user was replying to the topic/OP.

Now when replying to the topic, the author will be prioritized in the list.

All other cases are the same as before.

Added testing for all cases and changed `replyingToUser` to `replyingToUserId` for clarity and consistency with API.