PERF: Reuse existing core JS build where possible
Building the Discourse JS app is very resource-intensive. This commit introduces an `assemble_ember_build` script which will check the existing content of the `dist/` directory and re-use the core build if possible. Plugins will always be rebuilt.

For now, this functionality is only useful for multi-stage (i.e. non-standard) Discourse deployments. But in future, this script may be extended to pull the contents of the `dist/` directory from a remote location.