PERF: Compile main locale bundles just-in-time
Previously all locale bundles would be built & compressed during assets:precompile. For most sites, only one of these languages was actually used, so this is fairly wasteful.

This commit moves the main locale bundle into the ExtraLocalesController, which has recently undergone many improvements to make it more efficient. This allows locale files to be bundled "just in time" when they're first accessed.

The only material difference is in the compression level of the assets. In assets:precompile, we used a brotli compression level of 6. Now that they're generated dynamically, they'll use NGINX's on-the-fly compression. Assuming https://github.com/discourse/discourse/pull/32333 is merged, that means a brotli compression level of 4.

On the standard `en` bundle, that means the size will be increased from 117.1kB to 126.5kB.