DEV: unlink tempfiles when conversion fails
In the case that a conversion of an upload fails (e.g. ENOSPC when
writing to the directory) we were not unlinking the temporary file that
was being written to.

This modifies that way the we call into the #execute_convert method so
its consistent and the caller is able to clean up the file when an
exception has been raised.

See t/152709.