A11Y: site-text-logo should not be an H1 (#32405)
Without a site logo we render a the site name in the header as text,
this is a nice fallback!


![image](https://github.com/user-attachments/assets/007b9ad3-cbd7-4bcf-b5a3-6e4c813609b2)

...but it's marked up as an H1, and having the same H1 on every page
means it's not a meaningful page heading and can make headings fairly
annoying to navigate using screenreaders and other assistive
technologies

This converts it to a div and ensures that it retains the bold styling
previously applied by the heading tag. Font size is already handled by
the ID, so there are no visual changes as a result.