DEV: Update workflow concurrency controls to use `github.ref_name`
Not sure why we use `github.run_number` when it is a unique number for
each run of a workflow.

`github.head_ref` is only present on pull request events.

By updating the key to `github.ref_name` + `github.job` we ensure that
only one workflow will be running for a given PR or for a given branch.